requirements.txt
selenium>=4.39.0
requests>=2.32.5
robotframework>=7.4
robotframework-assertion-engine>=3.0.3
robotframework-browser>=19.12.3
robotframework-databaselibrary>=2.4.0
mysql-connector-python>=9.5.0
robotframework-pabot>=5.1.0
robotframework-pythonlibcore>=4.4.1
robotframework-requests>=0.9.7
robotframework-seleniumlibrary>=6.8.0
robotframework-stacktrace>=0.4.1
robotframework-datadriver>=1.11.2
cryptography>=46.0.3